fix: responsibility optimize
